# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Valheim Server UI API
  version: 1.0.0
  description: |
    HTTP API of valheim-server-ui. All endpoints are under /api/v1.
    Roles: viewer < operator < admin. Each operation declares the minimum role in `x-role`.
    Mutating requests must send header `X-Requested-With: valheim-ui` (CSRF guard).
    Long-running operations return 202 with a Job; follow progress via /events (SSE).
servers:
  - url: /api/v1
...
paths:
...
  /instances/{instanceId}/mods:
    parameters: [{ $ref: '#/components/parameters/instanceId' }]
    get:
      tags: [mods]
      x-role: viewer
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ModsOverview' }
    post:
      tags: [mods]
      summary: Install a Thunderstore package (and its dependencies)
      x-role: operator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [owner, name]
              properties:
                owner: { type: string }
                name: { type: string }
                version: { type: string, description: 'Omit for latest' }
      responses:
        '202': { $ref: '#/components/responses/JobResponse' }
        '404': { $ref: '#/components/responses/Error' }
  /instances/{instanceId}/mods/upload:
    post:
      tags: [mods]
      summary: Upload a mod zip (Thunderstore layout with manifest.json) or a single .dll
      x-role: operator
      parameters: [{ $ref: '#/components/parameters/instanceId' }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        '202': { $ref: '#/components/responses/JobResponse' }
        '422': { $ref: '#/components/responses/Error' }
  /instances/{instanceId}/mods/bepinex:
    parameters: [{ $ref: '#/components/parameters/instanceId' }]
    post:
      tags: [mods]
      summary: Install or upgrade BepInEx (instance must be stopped)
      x-role: operator
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StopIfRunning' }
      responses:
        '202': { $ref: '#/components/responses/JobResponse' }
    patch:
      tags: [mods]
      summary: Enable or disable the BepInEx loader
      x-role: operator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enabled]
              properties:
                enabled: { type: boolean }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ModsOverview' }
  /instances/{instanceId}/mods/{modId}:
    parameters:
      - { $ref: '#/components/parameters/instanceId' }
      - { $ref: '#/components/parameters/modId' }
    patch:
      tags: [mods]
      x-role: operator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enabled]
              properties:
                enabled: { type: boolean }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                required: [mod]
                properties:
                  mod: { $ref: '#/components/schemas/Mod' }
    delete:
      tags: [mods]
      x-role: operator
      responses:
        '202': { $ref: '#/components/responses/JobResponse' }
  /instances/{instanceId}/mods/{modId}/update:
    post:
      tags: [mods]
      summary: Update a Thunderstore mod to the latest (or given) version
      x-role: operator
      parameters:
        - { $ref: '#/components/parameters/instanceId' }
        - { $ref: '#/components/parameters/modId' }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                version: { type: string }
      responses:
        '202': { $ref: '#/components/responses/JobResponse' }
  /instances/{instanceId}/mods/configs:
    get:
      tags: [mods]
      x-role: viewer
      parameters: [{ $ref: '#/components/parameters/instanceId' }]
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                required: [files]
                properties:
                  files: { type: array, items: { $ref: '#/components/schemas/ConfigFileInfo' } }
  /instances/{instanceId}/mods/configs/{fileName}:
    parameters:
      - { $ref: '#/components/parameters/instanceId' }
      - in: path
        name: fileName
        required: true
        schema: { type: string }
    get:
      tags: [mods]
      x-role: operator
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ConfigFile' }
    put:
      tags: [mods]
      description: Send either `raw` (whole file) or `values` (key updates, comments preserved).
      x-role: operator
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ConfigFileUpdate' }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ConfigFile' }

  # ------------------------------------------------------------------ thunderstore
  /thunderstore/packages:
    get:
      tags: [thunderstore]
      x-role: viewer
      parameters:
        - { in: query, name: q, schema: { type: string } }
        - { in: query, name: category, schema: { type: string } }
        - { in: query, name: sort, schema: { type: string, enum: [rating, downloads, updated, name], default: rating } }
        - { in: query, name: include_deprecated, schema: { type: boolean, default: false } }
        - { in: query, name: page, schema: { type: integer, default: 1, minimum: 1 } }
        - { in: query, name: page_size, schema: { type: integer, default: 50, maximum: 100 } }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PackageSearchResult' }
  /thunderstore/packages/{owner}/{name}:
    get:
      tags: [thunderstore]
      x-role: viewer
      parameters:
        - { in: path, name: owner, required: true, schema: { type: string } }
        - { in: path, name: name, required: true, schema: { type: string } }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Package' }
        '404': { $ref: '#/components/responses/Error' }
...
components:
  parameters:
    instanceId: { in: path, name: instanceId, required: true, schema: { type: string, pattern: '^[a-z0-9][a-z0-9-]{0,31}$' } }
...
    modId: { in: path, name: modId, required: true, schema: { type: integer } }
...
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    JobResponse:
      description: Job
      content:
        application/json:
          schema:
            type: object
            required: [job]
            properties:
              job: { $ref: '#/components/schemas/Job' }
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: |
                One of: unauthorized, forbidden, not_found, validation_failed, conflict,
                instance_running, instance_not_installed, instance_busy, port_in_use,
                last_admin, invalid_credentials, account_locked, account_disabled,
                setup_done, oidc_disabled, oidc_error, steamcmd_missing, bepinex_missing,
                package_not_found, upstream_error, internal
            message: { type: string }
            details:
              type: object
              additionalProperties: true
              properties:
                fields:
                  type: array
                  items:
                    type: object
                    required: [field, message]
                    properties:
                      field: { type: string }
                      message: { type: string }
...
    StopIfRunning:
      type: object
      properties:
        stop_if_running: { type: boolean, default: false, description: Stop the instance for the operation and start it again afterwards }
...
    Mod:
      type: object
      required: [id, instance_id, source, owner, name, version, enabled, installed_at]
      properties:
        id: { type: integer }
        instance_id: { type: string }
        source: { type: string, enum: [thunderstore, manual] }
        owner: { type: string }
        name: { type: string }
        version: { type: string }
        enabled: { type: boolean }
        latest_version: { type: string, description: From the Thunderstore index, if known }
        update_available: { type: boolean }
        dependencies: { type: array, items: { type: string } }
        icon_url: { type: string }
        website_url: { type: string }
        installed_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    ModsOverview:
      type: object
      required: [bepinex, mods]
      properties:
        bepinex:
          type: object
          required: [installed, enabled]
          properties:
            installed: { type: boolean }
            enabled: { type: boolean }
            version: { type: string }
            latest_version: { type: string }
        mods: { type: array, items: { $ref: '#/components/schemas/Mod' } }
        pending_restart: { type: boolean }
    ConfigFileInfo:
      type: object
      required: [name, size_bytes, modified_at]
      properties:
        name: { type: string }
        size_bytes: { type: integer, format: int64 }
        modified_at: { type: string, format: date-time }
    ConfigEntry:
      type: object
      required: [section, key, value]
      properties:
        section: { type: string }
        key: { type: string }
        value: { type: string }
        description: { type: string }
        type: { type: string, description: 'From "# Setting type:" e.g. Boolean, Int32, Single, String, KeyCode, enum name' }
        default_value: { type: string }
        acceptable_values: { type: array, items: { type: string } }
        range: { type: object, properties: { min: { type: string }, max: { type: string } } }
    ConfigFile:
      type: object
      required: [name, raw, entries]
      properties:
        name: { type: string }
        raw: { type: string }
        entries: { type: array, items: { $ref: '#/components/schemas/ConfigEntry' } }
        modified_at: { type: string, format: date-time }
    ConfigFileUpdate:
      type: object
      properties:
        raw: { type: string }
        values:
          type: array
          items:
            type: object
            required: [section, key, value]
            properties:
              section: { type: string }
              key: { type: string }
              value: { type: string }
    PackageVersion:
      type: object
      required: [version, download_url, dependencies, date_created]
      properties:
        version: { type: string }
        description: { type: string }
        download_url: { type: string }
        dependencies: { type: array, items: { type: string } }
        downloads: { type: integer }
        file_size: { type: integer, format: int64 }
        date_created: { type: string, format: date-time }
    PackageSummary:
      type: object
      required: [owner, name, full_name, latest_version, rating_score, total_downloads, is_deprecated, categories, date_updated]
      properties:
        owner: { type: string }
        name: { type: string }
        full_name: { type: string }
        description: { type: string }
        icon_url: { type: string }
        package_url: { type: string }
        latest_version: { type: string }
        rating_score: { type: integer }
        total_downloads: { type: integer }
        is_deprecated: { type: boolean }
        is_pinned: { type: boolean }
        categories: { type: array, items: { type: string } }
        date_updated: { type: string, format: date-time }
    Package:
      allOf:
        - $ref: '#/components/schemas/PackageSummary'
        - type: object
          required: [versions]
          properties:
            website_url: { type: string }
            versions: { type: array, items: { $ref: '#/components/schemas/PackageVersion' } }
    PackageSearchResult:
      type: object
      required: [packages, total, page, page_size, index_updated_at]
      properties:
        packages: { type: array, items: { $ref: '#/components/schemas/PackageSummary' } }
        total: { type: integer }
        page: { type: integer }
        page_size: { type: integer }
        index_updated_at: { type: string, format: date-time }
    JobType:
      type: string
      enum: [install, update, backup, restore, world_import, world_regenerate, mod_install, mod_update, mod_uninstall, bepinex_install, scheduled_restart, thunderstore_refresh, self_upgrade]
    JobStatus:
      type: string
      enum: [queued, running, succeeded, failed, cancelled]
    Job:
      type: object
      required: [id, type, status, created_at]
      properties:
        id: { type: string, format: uuid }
        type: { $ref: '#/components/schemas/JobType' }
        instance_id: { type: string }
        status: { $ref: '#/components/schemas/JobStatus' }
        title: { type: string, description: Human readable, e.g. "Install Jotunn 2.29.2" }
        requested_by: { type: string }
        created_at: { type: string, format: date-time }
        started_at: { type: string, format: date-time }
        finished_at: { type: string, format: date-time }
        error: { type: string }
        summary:
          type: object
          additionalProperties: true
...